# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GDGOC Hacktoberfest Backend",
    "description": "Event-management API for GDGOC Hacktoberfest open-source contribution tracking.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/webhooks/github": {
      "post": {
        "tags": [
          "Webhooks"
        ],
        "summary": "GitHub Webhook Receiver",
        "description": "Primary GitHub Webhook receiver (v2 Architecture - No Celery / No Redis):\n1. Cryptographically verifies HMAC-SHA256 signature against GITHUB_WEBHOOK_SECRET.\n2. Persists payload to Table #11 (`webhook_jobs`) with status 'pending'.\n3. Executes event dispatch through the job engine:\n   - 'issues' (status/difficulty/category sync)\n   - 'pull_request' (auto-links PRs to claimed issues by GitHub username)\n   - 'push' (records commits and links to contributors/issues)\n   - 'pull_request_review' (updates contribution state and notifies student)\n4. Automatically handles retries with exponential backoff on transient errors.",
        "operationId": "github_webhook_receiver_webhooks_github_post",
        "parameters": [
          {
            "name": "X-GitHub-Event",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "ping",
              "title": "X-Github-Event"
            }
          },
          {
            "name": "X-Hub-Signature-256",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Hub-Signature-256"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/webhooks/jobs/drain": {
      "post": {
        "tags": [
          "Webhooks"
        ],
        "summary": "Drain Due Webhook Jobs (Supabase pg_cron Sweep)",
        "description": "Drains pending or due-for-retry rows from Table #11 (`webhook_jobs`).\nScheduled via Supabase pg_cron or invoked for manual queue sweeps.",
        "operationId": "drain_webhook_jobs_webhooks_jobs_drain_post",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/webhooks/jobs": {
      "get": {
        "tags": [
          "Webhooks"
        ],
        "summary": "List Webhook Jobs",
        "description": "List recent webhook jobs for monitoring and audit logging.",
        "operationId": "list_webhook_jobs_webhooks_jobs_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/users/me": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get own profile",
        "description": "Retrieve the full profile for the authenticated student/maintainer/admin.",
        "operationId": "get_my_profile_users_me_get",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Bearer JWT access token",
              "title": "Authorization"
            },
            "description": "Bearer JWT access token"
          },
          {
            "name": "X-User-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Test/Dev User ID header",
              "title": "X-User-Id"
            },
            "description": "Test/Dev User ID header"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Users"
        ],
        "summary": "Update own profile",
        "description": "Update profile details (such as display name or linked GitHub username) for the authenticated user.",
        "operationId": "update_my_profile_users_me_patch",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Bearer JWT access token",
              "title": "Authorization"
            },
            "description": "Bearer JWT access token"
          },
          {
            "name": "X-User-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Test/Dev User ID header",
              "title": "X-User-Id"
            },
            "description": "Test/Dev User ID header"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdateProfileRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/users/{user_id}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get public profile",
        "description": "Retrieve a public read-only contributor profile with contribution stats and recent open-source activity.",
        "operationId": "get_public_profile_users__user_id__get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPublicProfileResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ClaimUserBrief": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "psit_roll_no": {
            "type": "string",
            "title": "Psit Roll No"
          },
          "github_username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Github Username"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "psit_roll_no"
        ],
        "title": "ClaimUserBrief"
      },
...
      "ContributionPRBrief": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "github_pr_id": {
            "type": "integer",
            "title": "Github Pr Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "id",
          "github_pr_id",
          "title",
          "status"
        ],
        "title": "ContributionPRBrief"
      },
      "ContributionResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "issue_id": {
            "type": "integer",
            "title": "Issue Id"
          },
          "pr_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pr Id"
          },
          "status": {
            "$ref": "#/components/schemas/ContributionStatus"
          },
          "validation_status": {
            "$ref": "#/components/schemas/ContributionValidation"
          },
          "timeline_json": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Timeline Json"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "user": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ClaimUserBrief"
              },
              {
                "type": "null"
              }
            ]
          },
          "issue": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LinkedIssueBrief"
              },
              {
                "type": "null"
              }
            ]
          },
          "pull_request": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ContributionPRBrief"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "issue_id",
          "status",
          "validation_status",
          "created_at",
          "updated_at"
        ],
        "title": "ContributionResponse"
      },
      "ContributionStatus": {
        "type": "string",
        "enum": [
          "claimed",
          "in_progress",
          "pr_submitted",
          "under_review",
          "changes_requested",
          "accepted",
          "merged"
        ],
        "title": "ContributionStatus"
      },
      "ContributionValidation": {
        "type": "string",
        "enum": [
          "valid",
          "pending",
          "rejected",
          "duplicate",
          "invalid"
        ],
        "title": "ContributionValidation"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LinkedIssueBrief": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "github_issue_id": {
            "type": "integer",
            "title": "Github Issue Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "difficulty": {
            "type": "string",
            "title": "Difficulty"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "id",
          "github_issue_id",
          "title",
          "difficulty",
          "status"
        ],
        "title": "LinkedIssueBrief"
      },
...
      "UserProfileResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "psit_roll_no": {
            "type": "string",
            "title": "Psit Roll No"
          },
          "verified": {
            "type": "boolean",
            "title": "Verified"
          },
          "verified_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verified At"
          },
          "verification_method": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VerificationMethod"
              },
              {
                "type": "null"
              }
            ]
          },
          "github_username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Github Username"
          },
          "github_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Github Id"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "email",
          "psit_roll_no",
          "verified",
          "role",
          "created_at",
          "updated_at"
        ],
        "title": "UserProfileResponse"
      },
      "UserPublicProfileResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "github_username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Github Username"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "verified": {
            "type": "boolean",
            "title": "Verified"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "contributions_count": {
            "type": "integer",
            "title": "Contributions Count"
          },
          "merged_prs_count": {
            "type": "integer",
            "title": "Merged Prs Count"
          },
          "active_claims_count": {
            "type": "integer",
            "title": "Active Claims Count"
          },
          "recent_contributions": {
            "items": {
              "$ref": "#/components/schemas/ContributionResponse"
            },
            "type": "array",
            "title": "Recent Contributions",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "role",
          "verified",
          "created_at",
          "contributions_count",
          "merged_prs_count",
          "active_claims_count"
        ],
        "title": "UserPublicProfileResponse"
      },
      "UserRole": {
        "type": "string",
        "enum": [
          "student",
          "maintainer",
          "admin"
        ],
        "title": "UserRole"
      },
      "UserUpdateProfileRequest": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255,
                "minLength": 2
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "github_username": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Github Username"
          }
        },
        "type": "object",
        "title": "UserUpdateProfileRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VerificationMethod": {
        "type": "string",
        "enum": [
          "qr_auto",
          "manual"
        ],
        "title": "VerificationMethod"
      },
...
      "WebhookResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "event": {
            "type": "string",
            "title": "Event"
          },
          "action": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Action"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "status",
          "event",
          "detail"
        ],
        "title": "WebhookResponse"
      }
    }
  }
}